Rename opscan to semverscan
